+++ CHANGELOG.rst
@@ -20,7 +20,7 @@ Added
     - App alerts for sheet cache updates (#1000)
     - App alerts for zone owner for zone actions (#1204)
     - ``member_notify_move`` app setting (#1203)
-    - Project member notifications from zone moving (#1203)
+    - Project member notifications from zone moving (#1203, #1232)
     - ``LandingZone.user_message`` field (#1203)
 - **Samplesheets**
     - Simple link support for string cell rendering (#1001)